usaer_system/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.3
tzdata==2025.2
